requirements-dev.txt
pycodestyle
pytest
vcrpy
ansible_runner
mazer
git+https://github.com/SatelliteQE/nailgun.git@master#egg=nailgun
git+https://github.com/Apipie/apypie.git@master#egg=apypie
-r requirements.txt
